fix: handle delete wallet file request properly
