+++ desktop/main/Wallet.ts
@@ -32,29 +32,6 @@ const list = async () => {
 };
 
 //
-// FS Interactions
-//
-
-const deleteWalletFile = async (context: AppContext, filepath: string) => {
-  if (!context.mainWindow) return;
-  const options = {
-    title: 'Delete File',
-    message: 'All wallet data will be lost. Are You Sure?',
-    buttons: ['Delete Wallet File', 'Cancel'],
-  };
-  const { response } = await dialog.showMessageBox(context.mainWindow, options);
-  if (response === 0) {
-    try {
-      StoreService.clear();
-      await fs.unlink(filepath);
-      context.mainWindow.reload();
-    } catch (error) {
-      logger.error('deleteWalletFile', error);
-    }
-  }
-};
-
-//
 // Wallet data constructors
 // Pure
 //
@@ -159,7 +136,7 @@ export const createWallet = async ({
   return { path: walletPath, wallet, password };
 };
 
-const subscribe = (context: AppContext) => {
+const subscribe = () => {
   ipcMain.handle(ipcConsts.READ_WALLET_FILES, list);
 
   ipcMain.handle(ipcConsts.W_M_BACKUP_WALLET, (_event, filePath: string) =>
@@ -174,10 +151,6 @@ const subscribe = (context: AppContext) => {
     StoreService.set('walletFiles', newWalletFiles);
     return newWalletFiles;
   });
-
-  ipcMain.on(ipcConsts.W_M_SHOW_DELETE_FILE, (_event, filepath) =>
-    deleteWalletFile(context, filepath)
-  );
 };
 
 export default { subscribe };
