+++ desktop/main/startApp.ts
@@ -34,6 +34,7 @@ import handleOpenDashboard from './reactions/handleOpenDashboard';
 import { makeSubscription } from './rx.utils';
 import nodeIPCStreams, { nodeAndAppLogsListener } from './sources/node.ipc';
 import handleWipeOut from './reactions/wipeOut.ipc';
+import handleDeleteWalletFile from './reactions/deleteWalletFile.ipc';
 
 const loadNetworkData = () => {
   const $managers = new $.Subject<Managers>();
@@ -192,8 +193,9 @@ const startApp = (): AppStore => {
     handleSmesherIpc($managers, $smeshingStarted),
     // Handle show file
     handleShowFile($currentNetwork),
-    // Handle wipe out request
+    // IPC Reactions
     handleWipeOut($mainWindow, $isAppClosing),
+    handleDeleteWalletFile($mainWindow),
     // Push updates to Renderer process (redux via IPC)
     syncToRenderer(
       $mainWindow,
